examples/permission/requirement.txt
aiosqlite
graia-application-mirai
graia-broadcast